requirements.txt
# Nova: The Ambient Personal AI
# Core dependencies
sounddevice
Flask
requests
python-dotenv
rpi_ws281x
pytest
numpy
scipy
PyYAML
icalendar
Pillow
# Add more as modules are implemented
